edit types

+++ problems/1.intro/remove_duplicates/tests/test.rs
@@ -4,22 +4,19 @@ use remove_duplicates::remove_duplicates;
 fn test1() {
     let mut nums = vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
     let k = remove_duplicates(&mut nums);
-    assert_eq!(nums[..k as usize], vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
+    assert_eq!(nums[..k], vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
 }
 
 #[test]
 fn test2() {
     let mut nums = vec![1, 1, 2, 2, 2, 3, 4, 5, 6, 7, 8, 9, 10];
     let k = remove_duplicates(&mut nums);
-    assert_eq!(
-        nums[..k as usize],
-        vec![1, 1, 2, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-    );
+    assert_eq!(nums[..k], vec![1, 1, 2, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
 }
 
 #[test]
 fn test3() {
     let mut nums = vec![1, 2, 2, 3, 3];
     let k = remove_duplicates(&mut nums);
-    assert_eq!(nums[..k as usize], vec![1, 2, 2, 3, 3]);
+    assert_eq!(nums[..k], vec![1, 2, 2, 3, 3]);
 }
